Worked on diskdatabase. Implemented addNewsGroup(), addArticle, removeNewsGroup(), removeArticle(), initDatabase(), saveDBInfo(), getNewsgroupName(), getArticle(), newsgroupInDB() and finally listnewsgroups().
Reading to and from files is working, but there is a problem when running listNewsgroups() from the GUI-client. It
runs fine from within the database, i.e. on startup... but crashes the server when called from the GUI.

Curiouslyy, the GUI doesn't even seem to even ENTER the function, before it crashes.